infrastructure/tools/core/codecrawler/src/edu/uci/ics/sourcerer/codecrawler/util/StringFormatUtils.java
/*
 * Sourcerer: An infrastructure for large-scale source code analysis.
 * Copyright (C) by contributors. See CONTRIBUTORS.txt for full list.
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program. If not, see <http://www.gnu.org/licenses/>.
 * 
 */
package edu.uci.ics.sourcerer.codecrawler.util;

import java.text.ParseException;
import java.text.SimpleDateFormat;
import java.util.Date;

/**
 * @author <a href="mailto:hahrot@yahoo.com">Huy A. Huynh</a>
 * 
 */
public class StringFormatUtils {
	
	public static Date parseDateIn(String pattern, String string) {
		SimpleDateFormat dateFormat = new SimpleDateFormat(pattern);
		for (int i = 0; i < string.length()-6; i++) {		//well, at least 6 characters
			try {
				Date date = dateFormat.parse(string.substring(i));
				return date;
			} catch (ParseException e) {}
		}
		return null;
	}
	
	public static String formatDate(String pattern, Date date) {
		SimpleDateFormat dateFormat = new SimpleDateFormat(pattern);
		return dateFormat.format(date);
	}
	
}
